requirements.txt
django
djangorestframework
django-extensions
PyMySQL
drf-yasg
aiogram
python-dotenv
requests
browser-cookie3
